Split prepare stages
